+++ hack/sync-crds.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+crds=( appcatalogentries appcatalogs apps catalogs charts )
+
+for crd in "${crds[@]}"; do
+        curl -s "https://raw.githubusercontent.com/giantswarm/apiextensions/master/config/crd/application.giantswarm.io_${crd}.yaml" > "../pkg/crds/${crd}.yaml"
+done